requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==4.3.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-socketio==5.10.0